+++ tests/test_auth.py
@@ -1,5 +1,6 @@
 import pytest
-from backend.app import create_app, db
+from backend import create_app
+from backend.models import db
 
 
 @pytest.fixture
